2popitka

+++ src/exercise1.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 int main(int argc, char** argv) {
-    int array[10];
+    float array[10];
     float sum = 0;
     for(int i = 0; i < 10; ++i) {
-        scanf("%d", &array[i]);
+        scanf("%e", &array[i]);
         sum +=array[i];
  }
     float ot = sum/10;
